place lines of text properly
we now calculate the highest a glyph can be above the cursor, and the
lowest it can be below the cursor, and add these together to work out
how big a line of text should be. We also use that max bearing above the
cursor to calculate where the cursor itself should be, relative to the
corner of a text box.